tools/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
requests>=2.25.0
ffmpeg-python>=0.2.0